Remove content-parsing
