--- content-parsing/index.ts
@@ -1,2 +1,0 @@
-export * from "./ContentText"
-export * from "./ExpandableContentText"
